+++ xl2roefact/ldxml.py
@@ -1,0 +1,28 @@
+#!./.wenv_xl2roefact/bin/python3
+
+import pylightxl as xl
+
+
+def ldxml():
+    """ LDXML - modul de incarcare a facturii in sistemul ANAF E-Factura
+
+    Identification:
+        code-name: `ldxml`
+        copyright: (c) 2023 RENWare Software Systems
+        author: Petre Iordanescu (petre.iordanescu@gmail.com)
+
+    Specifications:
+        document: `110-SRE-api_to_roefact_requirements.md` section `Componenta xl2roefact`
+        INTRARI: fisier `f-XML`
+        IESIRI: raport cu validarea si identificatorul incarcarii
+    """
+
+    import pylightxl as xl
+
+    ... #TODO my code here
+    return None
+
+
+
+
+