+++ CHANGELOG.md
@@ -16,10 +16,9 @@
 ## 0.1 -#TODO wip...
 
 
--#TODO ---------- short planning board ---------------------------------------------------------
-* -ai exemplu de fact emisa catre Petrom in folder 100-ANA
-* -ai un exemplu complet si complet agnostic (trimis Gigi) de factura format XML si PDF tiparit ca sa faci: (1) incarcare XML (2) geenrare PDF (3) compararea variantelor si identificarea schemei XSD + document specificatii ANAF ref sistemul E-Factura (PDF trimis Liviu)
--------------------------------------------------------------------------------------------------
+- ---[ #TODO short planning board ]---------------------------------------------------------
+* ai un exemplu complet si complet agnostic (trimis Gigi) de factura format XML si PDF tiparit ca sa faci: (1) incarcare XML (2) geenrare PDF (3) compararea variantelor si identificarea schemei XSD + document specificatii ANAF ref sistemul E-Factura (PDF trimis Liviu)
+
 * -#NOTE link Swagger servicii web: `https://mfinante.gov.ro/web/efactura/informatii-tehnice`
 * -#NOTE link specif API incarcare fact: `https://mfinante.gov.ro/static/10/eFactura/upload.html#/EFacturaUpload/handleRequest`
 -------------------------------------------------------------------------------------------------
@@ -28,14 +27,97 @@
 
 
 
-### 0.1.7 #NOTE_wip... not_yet-defined (...)
+### 0.1.8 #TODO_future... `xl2roefact.RDINV` ???-[invoice header, invoice IDs, owner & partner?] (...)
 
-* wip...
+* -#NOTE_PLAN_tbd... RDINV module ...just read file and identify big zones:
+    * [x] DONE@`231108piu_b` set directories environment & open invoice
+    * [x] DONE@`231108piu_c` parametrized worksheet name containing invoice (received as parameter)
+    * [x] FIXED@`231108piu_d` problem: __xls binary format is not supported__
+    * [x] DONE@`231108piu_f` secured "unwanted crashes" and defaulted to first worksheet found when not specified #TODO up comment must be subject of documentation update
+    * [x] DONE@'231119piu_a' created `meta_info` key ref processed Excel file
+    * [x] DONE@`231119piu_b` invoice_items_area ready
+    * [x] DONE@`231119piu_a` written `invoice` dict to `f-JSON` file
+    * [ ] update documentation for all `rdinv` module
+    * invoice header
+        * [ ] invoice header - issuer (owner)
+        * [ ] invoice header - partner (supplier or customer) (#NOTE supplier will be good for `PayValidaBoa` to get suppliers invoices)
+        * [ ] invoice header - invoice identification (number, date, currency oth important as "non VAT payer")
+    * invoice footer
+    * ... consider to refactor `xl2roefact` to `xl2xml` (#FIXME not yet sure - must think more because name is very restrictive as what module does)
+    * [ ] left OPEN ISSUES on: `0.1.7` release
+
+* wip... last used 231126piu_a
 
 
 
 
 
+
+### 0.1.7 `xl2roefact.RDINV` invoice items & metadata (231126 h05:00)
+
+* **left OPEN ISSUES in this release** [piu@231126: marked on todo plan of `0.1.8` version]
+    * --> _file `xl2roefact/rdinv.py`, function `__mk_kv_invoice_items_area(...)`:_ `FIXME this will be identified in `invoice_header_area` ==> should be changed accordingly`
+    * --> _file `xl2roefact\invoice_files/_PLAN_model_test_factura_generat_anaf.xml`, line 114:_ `<cbc:ID>S</cbc:ID> #FIXME clarify.me_ pare a fi TIPUL PRODUSULUI: (S)erviciu sau ??? (P)rodus sau ???`
+
+
+* 231126piu_a item(s) line extensions == item VAT total, item VALUE total
+
+* 231125piu_c changed dir `excel_invoices/` with `invoice_files/`
+
+* 231125piu_b make a plan from `__model_test_factura_generat_anaf.xml` marking TODOs & DONE tags & renaming as `_PLAN_...idem...`
+
+* 231125piu_a final review, complete find of invoice relevant columns and write them in `invoice` dict:
+    * quantity (cbc:InvoicedQuantity),
+    * VAT percent (cbc:Percent),
+    * description (cbc:Name),
+    * uom (cbc:unitCode),
+    * unit price (cbc:PriceAmount),
+    * currency (c:currencyID)
+
+* 231224piu_a FIXED VAT calculation for lines which specify a VAT as 0% - USE CASE lines for other taxes like "acciza"... - still unsolved for very simplified invoices (see code line containing text "`_vat_percent` calculation should also")
+
+* 231123piu_c `rdinv` module, defined function `libutils.isnumber(a_string: str) -> bool` to test is a string is valid as any kind of number
+
+* 231123piu_b `rdinv` module, function to search for a list of string in a list items (suitable to identify useful / relevant invoice columns from Excel format) ==> **`__find_str_in_list(str_to_find: str, list_to_search: list)`**
+
+* 231123piu_a prepared function `__mk_kv_invoice_items_area(...)` to transform `invoice_items_area` in "canonical JSON format" (as kv pairs)
+
+* 231122piu_a  `invoice_items_area` & `meta_info` review, check and clean code
+
+* 231121piu_a final `invoice` dict / JSON data: moved all effective data under key `"excel_data"` (preparing so for final storable invoice "real keys")
+
+* 231119piu_b `invoice_items_area` set unknown rows header to `<current line number>.NOTE-<seq>`, where `seq` is an ordered sequence of letters (ie, resulting something like: `1.a, 1.b, ...`)
+\* 231119piu_a closed:
+    * [x] invoice `meta_info` area
+    * [x] a more defined, clean & clear `invoice_items_area`
+    * [x] add `last_processing_UTCtime` in `meta_info` dictionary
+    * [x] written `invoice` dict to a JSON (`f-JSON` file, see doc: `https://apitoroefact.renware.eu/commercial_agreement/110-SRE-api_to_roefact_requirements.html#vedere-de-ansamblu-a-solutiei`)
+
+* 231112piu_a refactor:
+    *  [x] `base_proc` to `xl2roefact`
+    *  [x] `BASE_PROC` to `xl2roefact`
+    *  [x] update documentation
+    *  [x] publish site
+
+* 231109piu_a installed module `pylightxl` in base environment (ie, in project root) to allow modules calling from root for future web components & updated requirements.txt
+
+* 231108piu_f RDINV (`rdinv.py`) secure for "unwanted crashes" errors (missing info) & clean code
+
+* 2811080iu_e install `mkdocstrings`, update `mkdocs.yml` and create a markdown document for xl2roefact component (did) not succeeded search for 'FIXME temporary disabled because errs' ...longer_string_...
+
+* 231108piu_d documented that `XLSX` is the only Excel file format supported (in `110-SRE-api_to_roefact_requirements.md`) & published site
+
+* 231108piu_c xl2roefact (`xl2roefact.py`) transformed in class & parametrized worksheet name containing invoice ==> **class BaseProc()**
+
+* 231108piu_b RDINV (`rdinv.py`) module set directories environment & open invoice
+
+* 231108piu_a consolidate decomposition (from `231107piu_c`) ==> directory `.../xl2roefact/modules/`
+
+* 231107piu_c transform decomposition (from `231107piu_b`) into complete Python modules (making directories for each)
+
+* 231107piu_b created module files according to decomposition (specs doc `110-SRE-api_to_roefact_requirements.md`)
+
+* 230107piu_a made a first DRAFT runnable CLI program with `Typer` (`https://typer.tiangolo.com/`) from `xls2xml.py` ==> `dist/xls2xml.exe`
 
 
 
@@ -49,16 +131,16 @@
 
 
 
-### 0.1.5 init component *BASE_PROC* for CLI application (231106 h06:30)
+### 0.1.5 init component *xl2roefact* for CLI application (231106 h06:30)
 
 * 231106piu_d published changed commercial documents
-* 231106piu_c get test invoices: from client, a RENware one, a 3rd party one ==> directory `base_proc/test_data_and_specs/`
+* 231106piu_c get test invoices: from client, a RENware one, a 3rd party one ==> directory `xl2roefact/test_data_and_specs/`
 * 231106piu_b updated `doc_src/110-SRE-payments_validation_board_requirements.md` ref system decomposition & list of components
-* 231106piu_a prepare environment for *BASE_PROC* component:
-    * [x] make a distinct Python-Windows environment in `base_proc/`
-    * [x] set as git ignore the intended new environment directory: `.wenv_base_proc/`
+* 231106piu_a prepare environment for *xl2roefact* component:
+    * [x] make a distinct Python-Windows environment in `xl2roefact/`
+    * [x] set as git ignore the intended new environment directory: `.wenv_xl2roefact/`
     * [x] install required modules in new environment (`pylightxl` core module)
-    * [x] create specific requirements file: `requirements_base_proc.txt`
+    * [x] create specific requirements file: `requirements_xl2roefact.txt`
 
 
 
@@ -66,7 +148,7 @@
 
 * 231105piu_c create system backbone as directories structure (and initialized min with a `README_xxx.md` file):
     * `web_dashb_app/` for component *WEB_DASHB* (as renaming actual `api_to_roefact_app/`)
-    * `base_proc/` for component *BASE_PROC*
+    * `xl2roefact/` for component *xl2roefact*
     * `data/` for component *SYSTEM_DB*
 * 231105piu_b revisit and review existing directories and refactored:
     * `sysInit/` --> `sys_init/`